fix: redirect page not found to localized page
