+++ src/utils/locale.ts
@@ -1,4 +1,4 @@
-import { i18n } from 'i18n.config';
+import { Locale, i18n } from 'i18n.config';
 
 export const getDefaultLocale = () => {
   return i18n.defaultLocale;
@@ -7,3 +7,7 @@ export const getDefaultLocale = () => {
 export const getAllLocales = () => {
   return i18n.locales;
 };
+
+export const setLocaleAttribute = (locale: Locale): void => {
+  document.documentElement.lang = locale;
+}
